Add some typemap to the Python swig interface.
This these new typemap, you can now do some basic stuff with histograms.
I will commit a demo for this soon...

The _cv.cpp and _highgui.cpp generated files are also updated. The new version
of the _highgui.cpp also correct a problem with the new signature of the
function icvInitAVIWriter()


git-svn-id: http://code.opencv.org/svn/opencv/trunk/opencv@169 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6