Another patch from "Phil" on the OpenCV Mailing list:
In tsysa.cpp, the relative path for the testdata directory should be two
subdirectory levels about the module location.  This is because the cvtest in
the src directory is a libtool wrapper.  The real binary is in the .libs
directory below the src directory.

Note: this don't works "as is" when you do the compilation in a separate
directory.


git-svn-id: http://code.opencv.org/svn/opencv/trunk/opencv@450 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6