More correct handling of the "int *value" parameter of cvCreateTrackbar().
Previous solution was just taking the Python int, and give it to the C
cvCreateTrackbar() function. This is not OK, because this generates a local
variable, and pass the address of the local variable to cvCreateTrackbar().
After that, the trackbar callback use this address and put in it the new
position of the cursor. And, because the local variable don't exists any more,
this cause many problems.

The proposed fix mallocate() a new int, and thus, the callback can change it
without problems.


git-svn-id: http://code.opencv.org/svn/opencv/trunk/opencv@958 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6